requirements.txt
beautifulsoup4
dash
ipykernel
ipywidgets
jupyter
jupyterlab
matplotlib
mysql-connector-python
numpy
pandas
plotly
prince
psycopg[binary]
pymongo
requests
scikit-learn
scipy
seaborn
sidetable
sqlalchemy
wquantiles
python-dotenv